AI_ML/requirements.txt
transformers

langchain

faiss-cpu

sentence-transformers

openai

arxiv

groq    